+++ Source/Scene/FrameRateMonitor.js
@@ -98,7 +98,7 @@ define([
         this._pauseCount = 0;
 
         var that = this;
-        this._preRenderRemoveListener = this._scene.preRender.addEventListener(function(scene, time) {
+        this._preUpdateRemoveListener = this._scene.preUpdate.addEventListener(function(scene, time) {
             update(that, time);
         });
 
@@ -274,7 +274,7 @@ define([
      * @see FrameRateMonitor#isDestroyed
      */
     FrameRateMonitor.prototype.destroy = function() {
-        this._preRenderRemoveListener();
+        this._preUpdateRemoveListener();
 
         if (defined(this._visibilityChangeRemoveListener)) {
             this._visibilityChangeRemoveListener();