+++ CHANGES.md
@@ -3,9 +3,12 @@ Change Log
 
 ### 1.42 - 2017-02-01
 
-* Added optional scene request render mode to reduce CPU usage
-    * `scene.requestRenderMode` enables a mode which will only request new render frames on changes to the scene, or when the simulation time change exceeds `scene.maximumRenderTimeChange`.
+* Breaking changes
+    * `scene.preRender` and `scene.postRender` events are called immediately before and after scene rendering only if the scene renders a frame. See `scene.requestRenderMode`. Use `scene.preUpdate` and `scene.postUpdate` for task that require regular updates.
+* Added optional scene request render mode to reduce CPU usage. [#6065](https://github.com/AnalyticalGraphicsInc/cesium/pull/6065)
+    * `scene.requestRenderMode` enables a mode which will only request new render frames on changes to the scene, or when the simulation time change exceeds `scene.maximumRenderTimeChange`. 
     * `scene.requestRender` will explicitly request a new render frame when in request render mode.
+    * Added `scene.preUpdate` and `scene.postUpdate` events that are raised before and after the scene updates respectively. The scene is always updated before executing a potential render.
 
 ### 1.41 - 2017-01-02